When building coordinates, treat the last oxygen specially.
The last oxygen of a protein can be placed by measuring the N-Ca-C-O
angle instead of the non-existant psi (N-Ca-C-N+1) angle. Thus, we have
added a measurement for this angle during dataset creation, and handle
this special case during structure generation.

Structures with C-terminal oxygens/hydroxyls are still not supported.