Improve ProteinVocabulary and batching to generate expected shapes.
Sequences are typically represented as sequences of integers. In this
representation, it's necessary to include a specific pad character (in
this case, 20). However, when using one-hot vectors, the zero vector can
be an implicit pad character. Thus, SidechainNet now supports both of
these representations for batching.

It may be necessary to modify downstream functions who still expect
sequences to be generated as integer sequences.