Fix for #15. Use explicit None check.
This issue was caused when a structure was generated from angles that were
equal to 0. In ResidueBuilder.build_sc, there was a check to see if the
previous torsional value was None. However, the check used
"if last_torsion" instead of the more correct "if last_torsion is not
None", since the value should be allowed to be 0
(even if this is not physical).